+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file. This projec
 **Added**
 
 -   The ability to exit the application via D-Bus.
+-   Return 1 exit code on translation error in CLI.
 
 ## [2.3.1](https://github.com/crow-translate/crow-translate/tree/2.3.1) (2020-01-18)